APP_商城-添加购物车页面快速点击"+"号商品一直在添加 APP_医师-医师详情界面没有评价却显示评价的数量 APP_商城，后台添加的多个商品规格在手机端商品详情界面未显示 APP_首页-健康宝典，文章分享给朋友、朋友圈后仍显示“泰晟健康”（推荐给好友、咨询记录同理） APP_商城-购物车-添加商品到库存不足时再次进入商品详情页面仍能添加购物车成功 APP-我的-我的订单，待收货页面可多次点击“确认收货”按钮 APP_我的-我的订单，后台设置了用户下单后超时xx分钟则订单自动取消，实际手机端未取消 APP_商城-商品列表，后台添加成功的商品不设置价格时，手机端显示null APP_健康宝典里的文章详情页面建议各个分类的字体显示一致 APP_我的-我的订单，订单详情页面订单备注未填写时去掉“请输入备注”的提示信息 APP_新用户登录后在我的积分里日期界面显示不全 APP_登录界面用户名输入框中建议去掉“用户名”提示信息 APP_后台添加的热门商品在手机端商城界面布局显示不合理 APP_我的-个人资料-修改密码，点击“去修改”，输入密码小于6位和大于16位均能修改成功

+++ app/src/main/java/com/taisheng/now/bussiness/bean/result/PostageResultBean.java
@@ -4,7 +4,7 @@ public class PostageResultBean {
 
     public String id;
     public String name;
-    public int money;
+    public double money;
     public String createTime;
     public String updateTime;
     public String status;
